+++ services/horizon/internal/db2/schema/migrations/53_add_trades_rounding_slippage.sql
@@ -1,7 +1,7 @@
 -- +migrate Up
 
 -- Add is_dust to trades table
-ALTER TABLE history_trades ADD rounding_slippage numeric(1000, 7);
+ALTER TABLE history_trades ADD rounding_slippage bigint;
 ALTER TABLE history_trades ADD base_reserves bigint;
 ALTER TABLE history_trades ADD counter_reserves bigint;
